requirements.txt
djangorestframework
Django
django-cleanup
ffmpeg
requests
pendulum
pydub
jinja2